Use folder attribute for host and dc module only
There is folder attribute only for vcenter_host and vcenter_datacenter
modules only, for other modules it shouldn't be used and caused issues.

With this PR, we fix the vcenter_vm module and others to return the
real REST object data, instead of error ALREADY_EXISTS in case of the
rerun of the module.

Signed-off-by: Ondra Machacek <omachace@redhat.com>